requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
ecdsa==0.17.0
fastapi==0.75.1
h11==0.13.0
httptools==0.4.0
idna==3.3
jose==1.0.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pytesseract==0.3.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
unicorn==1.0.3
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2
